Added model tags as columns to Chargeback reports
(cherry picked from commit ca6925c58d35d3b64ac820d65439581ba6da5b3f)

https://bugzilla.redhat.com/show_bug.cgi?id=1374377

+++ app/controllers/report_controller/reports/editor.rb
@@ -1780,7 +1780,7 @@ module ReportController::Reports::Editor
       f_len = fields.length
       for f_idx in 1..f_len # Go thru fields in reverse
         f_key = fields[f_len - f_idx].last
-        next if f_key.ends_with?(*CHARGEBACK_ALLOWED_FIELD_SUFFIXES)
+        next if f_key.ends_with?(*CHARGEBACK_ALLOWED_FIELD_SUFFIXES) || f_key.include?('managed')
         headers.delete(f_key)
         col_formats.delete(f_key)
         fields.delete_at(f_len - f_idx)
